CirculantEmbedding

Documentation for CirculantEmbedding.

CirculantEmbedding.GaussianProcess

CirculantEmbedding.LinearFilter

CirculantEmbedding.aliased_sdf

CirculantEmbedding.approx_cov

CirculantEmbedding.choose_circ_m

CirculantEmbedding.choose_nfreq

CirculantEmbedding.compute_lags

Statistics.var

CirculantEmbedding.GaussianProcess — Method

GaussianProcess(mean, Γ::Kernel, mesh::Mesh; pad=0)

Construct a Gaussian process with kernel Γ on a mesh mesh.

CirculantEmbedding.LinearFilter — Type

LinearFilter{Q,P,T}

A linear filter from a Q dimensional space to a P dimensional space, with eltype T.

CirculantEmbedding.aliased_sdf — Method

aliased_sdf(Γ::Kernel{D,P}, k, Δ; K = 3) where {D,P}

Compute the aliased sdf for a given kernel.

CirculantEmbedding.approx_cov — Method

approx_cov(Γ::Kernel{D,P}, lags) where {D,P}

Function to approximate the covariance from the sdf.

CirculantEmbedding.choose_circ_m — Method

choose_circ_m(n::Int, pad)

Choose the m used in circulant embedding.

CirculantEmbedding.choose_nfreq — Method

choose_nfreq(lag)

Chooses the maximum number of frequencies to use to approximate a given number of lags.

CirculantEmbedding.compute_lags — Method

compute_lags(mesh::CartesianGrid{D,T}, pad) where {D,T}

Compute lags required for circulant embedding on a given mesh.

Statistics.var — Method

var(k::Kernel{D,P}) where {D,P}

Compute the variance of a stationary kernel.